+++ src/Pages/HomePage/components/index.tsx
@@ -1,5 +1,4 @@
 //Empty file
-import * as React from "react";
 import Box from "@mui/material/Box";
 import { DataGrid, GridColDef, GridValueGetterParams } from "@mui/x-data-grid";
