+++ src/Pages/ProgressTrackingPage/ProgressTrackingPage.tsx
@@ -1,5 +1,5 @@
-import React, { FC } from "react";
 import Container from "@mui/material/Container";
+import { FC } from "react";
 import Grid from "@mui/material/Grid";
 import Paper from "@mui/material/Paper";
