+++ src/components/global/Layout/Layout.tsx
@@ -1,4 +1,4 @@
-import React, { FC, ReactNode } from "react";
+import { FC, ReactNode } from "react";
 import "./index.scss";
 import { Sidebar } from "../Sidebar/Sidebar";
